backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
sentence-transformers==2.2.2
chromadb==0.4.18
langchain==0.0.350
openai==1.3.7
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
celery==5.3.4
redis==5.0.1
flower==2.0.1
aiofiles==23.2.1
python-magic==0.4.27
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1